+++ internal/provider/password_resource_sdk.go
@@ -58,7 +58,7 @@ func (r *PasswordResourceModel) RefreshFromOperationsCreatePasswordResponseBody(
 		r.Renewable = types.BoolValue(resp.Renewable)
 		r.Replica = types.BoolValue(resp.Replica)
 		r.Role = types.StringValue(string(resp.Role))
-		r.TTLSeconds = types.Float64Value(resp.TTLSeconds)
+		r.TTLSeconds = types.Int64Value(resp.TTLSeconds)
 		r.Username = types.StringValue(resp.Username)
 	}
 
@@ -109,7 +109,7 @@ func (r *PasswordResourceModel) RefreshFromOperationsGetPasswordResponseBody(ctx
 		r.Renewable = types.BoolValue(resp.Renewable)
 		r.Replica = types.BoolValue(resp.Replica)
 		r.Role = types.StringValue(string(resp.Role))
-		r.TTLSeconds = types.Float64Value(resp.TTLSeconds)
+		r.TTLSeconds = types.Int64Value(resp.TTLSeconds)
 		r.Username = types.StringValue(resp.Username)
 	}
 
@@ -165,7 +165,7 @@ func (r *PasswordResourceModel) RefreshFromOperationsUpdatePasswordResponseBody(
 		r.Renewable = types.BoolValue(resp.Renewable)
 		r.Replica = types.BoolValue(resp.Replica)
 		r.Role = types.StringValue(string(resp.Role))
-		r.TTLSeconds = types.Float64Value(resp.TTLSeconds)
+		r.TTLSeconds = types.Int64Value(resp.TTLSeconds)
 		r.Username = types.StringValue(resp.Username)
 	}
